+++ src/blinker/__init__.py
@@ -1,3 +1,7 @@
+from __future__ import annotations
+
+import typing as t
+
 from blinker.base import ANY
 from blinker.base import NamedSignal
 from blinker.base import Namespace
@@ -16,4 +20,19 @@ __all__ = [
     "signal",
 ]
 
-__version__ = "1.7.0"
+
+def __getattr__(name: str) -> t.Any:
+    if name == "__version__":
+        import importlib.metadata
+        import warnings
+
+        warnings.warn(
+            "The '__version__' attribute is deprecated and will be removed in"
+            " Blinker 1.9.0. Use feature detection or"
+            " 'importlib.metadata.version(\"blinker\")' instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return importlib.metadata.version("blinker")
+
+    raise AttributeError(name)